Make sure num_public is int
Since num_public is calculated by dividing the combo number with the
number of test users, the result is a float. num_public is then used as
an index to an array and is expected to be used as an integer.

This while it works with python2, fails with python3. Fix the code to
ensure num_public is an integer.

+++ testhelper/testhelper.py
@@ -71,7 +71,7 @@ def get_mount_parameter(test_info, share, combonum):
     """
     if (combonum > get_total_mount_parameter_combinations(test_info)):
         assert False, "Invalid combination number"
-    num_public = combonum / len(test_info["test_users"])
+    num_public = int(combonum / len(test_info["test_users"]))
     num_users = combonum % len(test_info["test_users"])
     return gen_mount_params(
         test_info["public_interfaces"][num_public],
